Revert all insertion of nullifiers
It's not yet clear when nullifiers should be inserted. The logic got added
to assemble, prepare, and event handling in various forms, but all of them
seem a bit wrong. Pulling it out again to reconsider.

The querying and spending of nullifiers does seem correct, so that logic
remains where it is.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>