Remove PR template and auto-close action
Remove the PR template. This project is still in an early phase of development
primarily driven by the maintainers. A heavyweight PR template only adds
friction when making changes. As the project matures, and if the quality of
PRs becomes an issue, we can then adopt a standard template.

Change the "stale" workflow to never auto-close any issues/PRs, but to label
all stale issues/PRs after 30 days.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>