Add more deterministic error handling in plugin tests
Many of these tests still have notorious race conditions, but the
purely Sleep-based fixes and the use of the allowRegisterErrors flag
(which triggers the test to be non-deterministic depending on whether
errors occur or not) have potentially made it worse.

This reverts some of the current workarounds and adds a specific wait
condition for the error callback in TestTransportRegisterPartialSuccess.

If other intermittent failures pop up again, we can try new methods of
addressing them.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>