Add GitHub action to override SDK version in examples and update workflows to commit changes.
This commit introduces a new action for overriding the SDK version in example projects and updates the release workflows to include steps for checking out the repository code and committing the updated SDK version.

Additionally, it ensures that the SDK version is updated in the examples during the release process.

Signed-off-by: David Wertenteil <david.wertenteil@kaleido.io>